backend/rPPG-main/requirements.txt
einops==0.7.0
h5py==2.10.0
yacs==0.1.8
scipy==1.5.2
pandas==1.1.5
scikit_image==0.17.2
numpy==1.22.0
matplotlib==3.1.2
opencv_python==4.5.2.54
PyYAML==6.0
scikit_learn==1.0.2
tensorboardX==2.4.1
timm==0.9.16
tqdm==4.66.3
mat73==0.59
ipykernel==6.26.0
ipywidgets==8.1.1
retina-face==0.0.13
# causal-conv1d==1.4.0
# mamba-ssm==2.2.2
thop==0.1.1.post2209072238